rename: move types folder to src path
