+++ src/utils/constants/filterConstants.ts
@@ -1,4 +1,4 @@
-import { FilterType } from '../../context/types/TodoTypes';
+import { FilterType } from '../../types/todoTypes';
 
 export const FILTERS: Array<{ label: string; value: FilterType }> = [
   { label: 'All', value: 'all' },
